+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - Print: Prevent red background at 400 dpi by forcing tiledWMS for large ImageWMS layers
 - FeatureLister: Use configured parameter `featurePrefix` if configured instead of fallback.
 - Legend: Removed fallback legend as it caused problems with other legends.
+- BufferAnalysis: Fixed applying values from URL. New and legacy format are now correctly parsed and applied.
 
 ---